Adds classes for image and video descriptors.
Added abstract class ImageDescriptor, parent for all image descriptors. All
image descriptors should contain a timestamp indicating the time in the video
where their corresponding frame was, and be able to be written into a byte array.

Video descriptors are parametrized by their image descriptor type, each video
descriptor contains an arraylist of it's corresponding image descriptors and it's
able to add new descriptors to it on demand.