Changed Keyframe descriptor to crop image from camera before computing descriptor.
The extractor now crops the image before sending it to the extraction rs script.
The cropping is hardcoded to take only the right half of the screen because its meant
as a proof of concept, the next step will be to take user input into account to define
the cropping boundaries.